[Modify] Rename project to better match standard library names [Add] Separate implementation of Range [Modify] README now include new namespace name and new implementation
