+++ Range/src/Range/Range.cs
@@ -1,6 +1,6 @@
-﻿using System.Linq;
+﻿using System;
 
-namespace System.Numerics.Range
+namespace System.Numerics
 {
     public class Range<T> where T : IComparable<T>
     {
@@ -29,39 +29,6 @@ namespace System.Numerics.Range
                 return this.Ceiling;
             else
                 return value;
-        }
-    }
-
-    public class LinearDiscreteRange : Range<decimal>
-    {
-        public LinearDiscreteRange(decimal floor, decimal ceiling, decimal interval) : base(floor, ceiling)
-        {
-            this.Interval = interval;
-        }
-        public readonly decimal Interval;
-
-        public override decimal FindClosestValue(decimal value)
-        {
-            if (value.CompareTo(this.Floor) < 0)
-                return this.Floor;
-            else if (value.CompareTo(this.Ceiling) > 0)
-                return this.Ceiling;
-            else
-                return Math.Abs((value - this.Floor) / this.Interval) * this.Interval;
-        }
-    }
-
-    public class DiscreteRange<T> : Range<T> where T : IComparable<T>
-    {
-        public DiscreteRange(IOrderedEnumerable<T> discreteValue) : base(discreteValue.First(), discreteValue.Last())
-        {
-            this.DiscreteValues = discreteValue;
-        }
-        public readonly IOrderedEnumerable<T> DiscreteValues;
-        
-        public override T FindClosestValue(T value)
-        {
-            return this.DiscreteValues.OrderBy(discreteValue => (uint)(value.CompareTo(discreteValue))).First();
         }
     }
 }
